History working as a plugin
